backend/requirements.txt
# FastAPI and async
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.12.1
pgvector==0.2.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Redis and Celery
redis==5.0.1
celery==5.3.4

# AI and NLP
openai==1.3.7
spacy==3.7.2
sentence-transformers==2.2.2

# File processing
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
numpy==1.24.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0